cs6263_template/src/myproject/requirements.txt
wandb
vllm
faiss-cpu
datasets
numpy
pytorch-lightning
pyyaml
sentence-transformers
torch
tqdm
transformers
pydantic
